Created SCC role
The SCAP-Compliance-Checker will scan the system based on the DISA RHEL7 STIG Benchmark and output XCCDF files to view in STIG Viewer and HTML reports to see compliance percentage and details of all STIG items.  The SCC rpm needs to be downloaded from the DISA website. The latest RHEL7 Benchmark needs to be downloaded as the zip file and moved over to the /opt/scc directory along with the SCC rpm.